Cleaned up flight screen UI
When there were <8 planes, the Ui was looking rough, now place holders are put in to make it look clean that are dynamically taken away when more planes are bought. Also introduced a new button that shows the plane's location when clicked. Fixed coin position bug and arraylist  bug where the same type of plane was being referenced to only one variable. Now each plane is independent of each other.